requirement.txt
streamlit
google-generativeai
python-dotenv
google-api-python-client
pydantic